Wire up CreateWorkout with Setup stage
